+++ src/main.cpp
@@ -15,6 +15,7 @@
 #include "capability_managment.hpp"
 #include "configuration.hpp"
 #include "crazytrace.hpp"
+#include "landlock.hpp"
 #include "nodecontainer.hpp"
 #include "seccomp.hpp"
 #include "tun_tap.hpp"
@@ -28,6 +29,19 @@ int main(int argc, char * argv[])
         CapabilityManagment::lock();
         CapabilityManagment::drop_capabilies();
 #endif
+#ifdef HAVE_LANDLOCK
+        const LandlockRuleset landlock_ruleset_init(
+            LANDLOCK_ACCESS_FS_READ_DIR | LANDLOCK_ACCESS_FS_REMOVE_DIR |
+                LANDLOCK_ACCESS_FS_REMOVE_FILE | LANDLOCK_ACCESS_FS_MAKE_CHAR |
+                LANDLOCK_ACCESS_FS_MAKE_DIR | LANDLOCK_ACCESS_FS_MAKE_REG |
+                LANDLOCK_ACCESS_FS_MAKE_SOCK | LANDLOCK_ACCESS_FS_MAKE_FIFO |
+                LANDLOCK_ACCESS_FS_MAKE_BLOCK | LANDLOCK_ACCESS_FS_MAKE_SYM |
+                LANDLOCK_ACCESS_FS_REFER,
+            LANDLOCK_ACCESS_NET_BIND_TCP | LANDLOCK_ACCESS_NET_CONNECT_TCP,
+            LANDLOCK_SCOPE_ABSTRACT_UNIX_SOCKET);
+        landlock_ruleset_init.restrict_self();
+#endif
+
 #ifdef HAVE_SECCOMP
         SeccompFilterContext seccomp_context(SCMP_ACT_ALLOW);
         // see also
@@ -49,7 +63,6 @@ int main(int argc, char * argv[])
         seccomp_context.kill_reboot();
         seccomp_context.kill_resources();
         seccomp_context.kill_setuid();
-        seccomp_context.kill_signal();
         seccomp_context.kill_swap();
         seccomp_context.kill_sync();
         seccomp_context.kill_system_service();
@@ -64,19 +77,38 @@ int main(int argc, char * argv[])
         const Configuration config(filename);
         config.get_log_level().apply();
 
-        BOOST_LOG_TRIVIAL(info) << "libtuntap version: " << TUNTAP_VERSION_MAJOR
-                                << "." << TUNTAP_VERSION_MINOR;
+        BOOST_LOG_TRIVIAL(info)
+            << "libtuntap version (compile time): " << TUNTAP_VERSION_MAJOR
+            << "." << TUNTAP_VERSION_MINOR;
         const int version = ::tuntap_version();
         const int major = (version >> 8) & 0xFF;
         const int minor = version & 0xFF;
         BOOST_LOG_TRIVIAL(info)
-            << "libtuntap version: " << major << "." << minor;
+            << "libtuntap version (runtime): " << major << "." << minor;
 
 #if defined(TINS_VERSION_MAJOR) && defined(TINS_VERSION_MINOR) && \
     defined(TINS_VERSION_PATCH)
         BOOST_LOG_TRIVIAL(info)
             << "libtins version: " << TINS_VERSION_MAJOR << "."
             << TINS_VERSION_MINOR << "." << TINS_VERSION_PATCH;
+#endif
+
+#ifdef HAVE_LIBCAPNG
+        BOOST_LOG_TRIVIAL(info) << "libcapng: true";
+#else
+        BOOST_LOG_TRIVIAL(info) << "libcapng: false";
+#endif
+
+#ifdef HAVE_SECCOMP
+        BOOST_LOG_TRIVIAL(info) << "seccomp: true";
+#else
+        BOOST_LOG_TRIVIAL(info) << "seccomp: false";
+#endif
+
+#ifdef HAVE_LANDLOCK
+        BOOST_LOG_TRIVIAL(info) << "landlock: true";
+#else
+        BOOST_LOG_TRIVIAL(info) << "landlock: false";
 #endif
 
         const std::shared_ptr<NodeContainer> nodecontainer =
@@ -103,6 +135,29 @@ int main(int argc, char * argv[])
 
 #ifdef HAVE_LIBCAPNG
         CapabilityManagment::drop_all_capabilies();
+#endif
+#ifdef HAVE_LANDLOCK
+        const LandlockRuleset landlock_ruleset_loop(
+            LANDLOCK_ACCESS_FS_EXECUTE | LANDLOCK_ACCESS_FS_WRITE_FILE |
+                LANDLOCK_ACCESS_FS_READ_FILE | LANDLOCK_ACCESS_FS_TRUNCATE |
+                LANDLOCK_ACCESS_FS_READ_DIR | LANDLOCK_ACCESS_FS_REMOVE_DIR |
+                LANDLOCK_ACCESS_FS_REMOVE_FILE | LANDLOCK_ACCESS_FS_MAKE_CHAR |
+                LANDLOCK_ACCESS_FS_MAKE_DIR | LANDLOCK_ACCESS_FS_MAKE_REG |
+                LANDLOCK_ACCESS_FS_MAKE_SOCK | LANDLOCK_ACCESS_FS_MAKE_FIFO |
+                LANDLOCK_ACCESS_FS_MAKE_BLOCK | LANDLOCK_ACCESS_FS_MAKE_SYM |
+                LANDLOCK_ACCESS_FS_REFER | LANDLOCK_ACCESS_FS_IOCTL_DEV,
+            LANDLOCK_ACCESS_NET_BIND_TCP | LANDLOCK_ACCESS_NET_CONNECT_TCP,
+            LANDLOCK_SCOPE_ABSTRACT_UNIX_SOCKET | LANDLOCK_SCOPE_SIGNAL);
+        landlock_ruleset_loop.add_path_beneath_rule(
+            LANDLOCK_ACCESS_FS_WRITE_FILE | LANDLOCK_ACCESS_FS_READ_FILE |
+                LANDLOCK_ACCESS_FS_IOCTL_DEV,
+            dev.native_handler());
+        landlock_ruleset_loop.restrict_self();
+#endif
+#ifdef HAVE_SECCOMP
+        seccomp_context.kill_signal();
+        seccomp_context.load();
+        seccomp_context.release();
 #endif
 
         const Crazytrace ct(