+++ src/seccomp.hpp
@@ -8,37 +8,39 @@
 #ifdef HAVE_SECCOMP
 
     #include <stdexcept>
+    #include <cstdint>
     #include <seccomp.h>
+    #include <sys/syscall.h>
 
 class SeccompFilterContext
 {
     public:
         explicit SeccompFilterContext(uint32_t def_action);
-        void rule_add(uint32_t action, int syscall);
-        void allow(int syscall);
-        void kill(int syscall);
-        void kill_chown();
-        void kill_clock();
-        void kill_cpu_emulation();
-        void kill_debug();
-        void kill_others();
-        void kill_ipc();
-        void kill_keyring();
-        void kill_memlock();
-        void kill_module();
-        void kill_mount();
-        void kill_obsolete();
-        void kill_privileged();
-        void kill_rawio();
-        void kill_reboot();
-        void kill_resources();
-        void kill_setuid();
-        void kill_signal();
-        void kill_swap();
-        void kill_sync();
-        void kill_system_service();
-        void load();
-        void reset(uint32_t def_action);
+        void rule_add(uint32_t action, int syscall) const;
+        void allow(int syscall) const;
+        void kill(int syscall) const;
+        void kill_chown() const;
+        void kill_clock() const;
+        void kill_cpu_emulation() const;
+        void kill_debug() const;
+        void kill_others() const;
+        void kill_ipc() const;
+        void kill_keyring() const;
+        void kill_memlock() const;
+        void kill_module() const;
+        void kill_mount() const;
+        void kill_obsolete() const;
+        void kill_privileged() const;
+        void kill_rawio() const;
+        void kill_reboot() const;
+        void kill_resources() const;
+        void kill_setuid() const;
+        void kill_signal() const;
+        void kill_swap() const;
+        void kill_sync() const;
+        void kill_system_service() const;
+        void load() const;
+        void reset(uint32_t def_action) const;
         void release();
         [[nodiscard]] bool is_useable() const noexcept;